Initialized stats tracking object properly.

+++ CompeteInstSet/source/actions/PopulationActions.cc
@@ -314,7 +314,7 @@ private:
   double m_merit;
   int m_lineage_label;
   double m_neutral_metric;
-  double m_inst_set_id;
+  int m_inst_set_id;
 public:
   cActionInjectSequence(cWorld* world, const cString& args)
   : cAction(world, args), m_cell_start(0), m_cell_end(-1), m_merit(-1), m_lineage_label(0), m_neutral_metric(0), m_inst_set_id(0)
